+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 ### BioBlend v0.14.0 - unreleased
 
 * Dropped support Python 2.7. Dropped support for Galaxy releases 14.10-17.05.
-  Added support for Python 3.8. Added support for Galaxy releases 19.09
-  and 20.01.
+  Added support for Python 3.8. Added support for Galaxy releases 19.09, 20.01
+  and 20.05.
 
 * Removed deprecated ``show_stderr()`` and ``show_stdout`` methods of
   ``DatasetClient``.